requirements.txt
tensorflow-cpu
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu
pandas
numpy
matplotlib
seaborn
scikit-learn
streamlit
ultralytics
opencv-python
gdown
